Successfully moved the right-side component from Recipes.js to MainRecipe.js. The snackbar works for delete and the SideRecipes update on a delete. Have not check the add recipe logic yet to see if I need to fix that too.
