+++ client/src/deleteRecipe.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import axios from "axios"
+
+export const deleteRecipe = () => {
+   return (
+1
+   )
+}
